Port forward: [1/3] Install persistence support
Copy files from OUT/install to ZIP/install and extract to /tmp/install
at install time.  This is an extensible framework that may be used for
other install tasks as needed.  Note that none of the install files are
only extracted to a ramdisk, they are never written to /system.
Run script persist.sh with argument "save" prior to formatting /system
and again with argument "restore" after /system is installed.  This may
be explicitly disabled by setting "TARGET_DISABLE_PERSIST := true".

Author: Tom Marshall

Signed-off-by: igio90 <giovanni.rocca.90@gmail.com>